graphics: Rebuild vertex indices from obj
OpenGL, unlike OBJ, doesn't allow separate index for normals and other
attributes. So, to avoid duplicating shared vertices, model data must be
rearranged.

+++ include/aw/graphics/gl/model.h
@@ -68,7 +68,7 @@ struct vert_data {
 };
 
 struct mesh_data {
-	array_view<GLushort> triangles;
+	array_view<GLuint> triangles;
 	array_view<model::subobject> list;
 	// TODO: materials
 };
